Rafting In India
Rafting, the high-octane activity of navigating a river in an inflatable raft, has multiple levels of difficulty depending on the river's choppyness. These 'difficulty levels' are determined by the presence of rapids, which form as a result of dramatic drops in the river's height, as well as the presence of small or large boulders in the water. Rafting is a difficult but extremely enjoyable exercise; just remember to follow the instructor's safety instructions! While whitewater (rapids) can be frightening, river running done properly – under professional supervision, with the proper training, using the appropriate equipment, taking all necessary safety precautions, and adhering to a set of international safety and environmental standards – can be a very safe, enjoyable, and exciting soft-adventure sport.